fix: break html parser import cycle
