+++ apps/worker/app/services/document_parser/formats/html/parser.py
@@ -446,127 +446,3 @@ def first_cols_rows_html(html_str, max_items=10, max_chars=20):
     first_col_text = " | ".join(unique_col) if unique_col else ""
 
     return first_row_text, first_col_text
-
-
-# HTML Document Parser — converts .html/.htm files into text lines
-# compatible with parse_md(), reusing the markdown processing pipeline.
-# Uses markdownify (already a transitive dependency via markitdown) for
-# robust HTML-to-Markdown conversion with proper DOM-order traversal.
-
-
-def parse_html(
-    output_dir: str,
-    source_type: str,
-    file_path: str,
-    base_llm_paras=None,
-    relative_root: str | None = None,
-):
-    """Parse an HTML file into a hierarchical document DataFrame.
-
-    Converts HTML to clean Markdown via markdownify, stripping unsafe and
-    non-content elements (script, style, img, iframe, etc.), then delegates
-    to the standard parse_md() pipeline for heading detection, hierarchy
-    reconstruction, table extraction, and LLM-based enrichment.
-
-    Args:
-        output_dir: Full output directory path for parsed artifacts.
-        source_type: Source type label (always "html" from the adapter).
-        file_path: Absolute path to the .html or .htm file on disk.
-        base_llm_paras: LLM parameter dict for summary enrichment.
-        relative_root: Root path segment for hierarchical path construction.
-
-    Returns:
-        pd.DataFrame with columns [path, content, type, summary, keywords]
-        representing the parsed document hierarchy.
-    """
-    from app.services.common.file_loading import load_file_bytes
-    from app.services.document_parser.formats.markdown.parser import parse_md
-
-    html_bytes = load_file_bytes(file_path, file_url="")
-    md_lines = _html_to_md_lines(html_bytes)
-
-    # parse_md() already calls merge_html_tables() internally, so we
-    # don't need to do it here — just pass the lines straight through.
-    parsed_df = parse_md(
-        output_dir,
-        source_type=source_type,
-        md_lines=md_lines,
-        base_llm_paras=base_llm_paras,
-        relative_root=relative_root,
-    )
-    return parsed_df
-
-
-def _html_to_md_lines(html_content: str | bytes) -> list[str]:
-    """Convert HTML document body into markdown-like text lines.
-
-    Uses markdownify with a custom converter subclass that:
-    - Strips unsafe/non-content tags (script, style, iframe, etc.)
-    - Converts <img> to alt-text-only (no broken markdown image refs)
-    - Preserves headings, lists, tables, code blocks, blockquotes
-    - Correctly handles inline formatting in DOM order
-
-    Returns a list of lines suitable for the parse_md() pipeline.
-    """
-    converter = _SafeHtmlConverter(
-        heading_style="ATX",
-        bullets="-",
-        strong_em_symbol="*",
-    )
-    md_text = converter.convert(html_content)
-
-    # Split into lines and strip trailing blanks.
-    lines = md_text.splitlines()
-    while lines and not lines[-1].strip():
-        lines.pop()
-
-    return lines
-
-
-class _SafeHtmlConverter:
-    """Markdownify-based HTML→Markdown converter with safety filtering.
-
-    Subclasses MarkdownConverter to suppress elements that are either
-    non-content (script, style, nav) or would produce broken references
-    in a single-file upload context (img with relative src).
-    """
-
-    def __new__(cls, **kwargs):
-        # Lazy import so the module loads even if markdownify is somehow
-        # missing — the error surfaces only when parse_html() is called.
-        from markdownify import MarkdownConverter
-
-        # Dynamically build the subclass from MarkdownConverter so we
-        # don't need a module-level import.
-        converter_cls = type(
-            "_SafeHtmlConverterImpl",
-            (MarkdownConverter,),
-            {
-                "convert_script": _noop_converter,
-                "convert_style": _noop_converter,
-                "convert_iframe": _noop_converter,
-                "convert_object": _noop_converter,
-                "convert_embed": _noop_converter,
-                "convert_form": _noop_converter,
-                "convert_nav": _noop_converter,
-                "convert_noscript": _noop_converter,
-                "convert_img": _img_alt_only_converter,
-            },
-        )
-        return converter_cls(**kwargs)
-
-
-def _noop_converter(self, el, text, parent_tags):
-    """Return empty string — suppresses the element and all its content."""
-    return ""
-
-
-def _img_alt_only_converter(self, el, text, parent_tags):
-    """Convert <img> to alt-text-only, avoiding broken markdown image refs.
-
-    Single-file HTML uploads don't have access to relative image assets,
-    and fetching remote URLs would create SSRF concerns. We preserve the
-    alt text as a bracketed annotation so the content isn't silently lost.
-    """
-    alt = el.get("alt", "").strip()
-    return f"[Image: {alt}]" if alt else ""
